[Refactor] 에러처리 함수 분리
